A couple of renamings to simplify things
