--- workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/Repo.kt
@@ -1,6 +1,0 @@
-package com.novoda.workshop
-
-internal data class Repo(
-    val id: Long,
-    val name: String
-)
