+++ workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/contributors/ContributorHolder.kt
@@ -6,7 +6,6 @@ import android.widget.ImageView
 import com.bumptech.glide.Glide
 import com.bumptech.glide.load.model.GlideUrl
 import com.novoda.workshop.R
-import com.novoda.workshop.User
 import kotlinx.android.extensions.LayoutContainer
 import kotlinx.android.synthetic.main.item_contributor.*
 
@@ -15,13 +14,13 @@ internal class ContributorHolder(override val containerView: View) : RecyclerVie
 
     private val resources = containerView.resources
 
-    fun bind(user: User) {
-        contributor_label.text = labelFrom(user)
-        contributor_image.loadImageFrom(user.avatarUrl)
+    fun bind(contributor: Contributor) {
+        contributor_label.text = labelFrom(contributor)
+        contributor_image.loadImageFrom(contributor.avatarUrl)
     }
 
-    private fun labelFrom(user: User) =
-        resources.getString(R.string.contributor_item_label, user.login, user.contributions)
+    private fun labelFrom(contributor: Contributor) =
+        resources.getString(R.string.contributor_item_label, contributor.name, contributor.contributions)
 
 }
