+++ workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/contributors/ContributorsAdapter.kt
@@ -4,11 +4,10 @@ import android.support.v7.widget.RecyclerView
 import android.view.LayoutInflater
 import android.view.ViewGroup
 import com.novoda.workshop.R
-import com.novoda.workshop.User
 
 internal class ContributorsAdapter : RecyclerView.Adapter<ContributorHolder>() {
 
-    private val contributors = mutableListOf<User>()
+    private val contributors = mutableListOf<Contributor>()
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ContributorHolder {
         val itemView = LayoutInflater.from(parent.context).inflate(R.layout.item_contributor, parent, false)
@@ -21,7 +20,7 @@ internal class ContributorsAdapter : RecyclerView.Adapter<ContributorHolder>() {
 
     override fun getItemCount(): Int = contributors.size
 
-    fun setContributors(contributors: List<User>) {
+    fun setContributors(contributors: List<Contributor>) {
         this.contributors.clear()
         this.contributors.addAll(contributors)
         notifyDataSetChanged()
