+++ workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/contributors/ContributorsBackend.kt
@@ -1,17 +1,15 @@
 package com.novoda.workshop.contributors
 
-import com.novoda.workshop.Repo
-import com.novoda.workshop.User
 import kotlinx.coroutines.Deferred
 import retrofit2.http.GET
 import retrofit2.http.Path
 
 internal interface ContributorsBackend {
     @GET("orgs/novoda/repos?per_page=100")
-    fun listRepos(): Deferred<List<Repo>>
+    fun listRepos(): Deferred<List<Repository>>
 
     @GET("repos/novoda/{repo}/contributors?per_page=100")
     fun listContributors(
         @Path("repo") repo: String
-    ): Deferred<List<User>>
+    ): Deferred<List<Contributor>>
 }
