+++ workshops/AndroidKotlinWorkshop/app/src/main/java/com/novoda/workshop/contributors/Repository.kt
@@ -1,0 +1,6 @@
+package com.novoda.workshop.contributors
+
+internal data class Repository(
+    val id: Long,
+    val name: String
+)
